requirements.txt
PyQt6==6.5.2
python-dotenv==1.0.0
google-generativeai==0.1.0rc3
openai==1.3.5
